# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Time control API",
        "description": "Este API permite gestionar y contabilizar el tiempo que dedicamos a proyectos",
        "contact": {
            "name": "Soporte API",
            "url": "https://escuela.it/contacto",
            "email": "contacto@escuela.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login de usuarios",
                "description": "Ruta para realizar el login de un usuario existente. Verifica el usuario y devuelve el token",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Credenciales de usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthResponse"
                    },
                    "401": {
                        "description": "Error de validación de la entrada de datos o usuario desconocido"
                    }
                }
            }
        },
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Registro de usuarios",
                "description": "Ruta para realizar el alta de un usuario. Crea el usuario y devuelve el token",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Objeto de usuario a crear",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/AuthResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationErrorResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del usuario. Al crear el usuario no se debe enviar el identificador, pues se generará en la creación.",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Clave del usuario",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "ValidationErrorResponse": {
                "description": "Respuesta cuando se ha encontrado un error de validación",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Validation error message",
                                    "type": "string"
                                },
                                "errors": {
                                    "description": "Detalles específicos sobre los errores de validación.",
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AuthResponse": {
                "description": "Usuario autenticado, se devuelve el token",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Mensaje de respuesta",
                                    "type": "string"
                                },
                                "token": {
                                    "description": "Token de autenticación",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}